fix(brick-calculator): change thickness selection value parser from parseInt to parseFloat to support decimal 4.5-inch partition wall

+++ src/tools/BrickCalculator.tsx
@@ -158,7 +158,7 @@ Estimated u/s standard brick joint ratios.`;
             </label>
             <select
               value={thickness}
-              onChange={(e) => setThickness(parseInt(e.target.value) || 9)}
+              onChange={(e) => setThickness(parseFloat(e.target.value) || 9)}
               className="saas-select animate-none"
             >
               <option value={9}>9-inch (Outer Wall)</option>
